+++ src/main/javascript/Core/AppEvents.js
@@ -1,0 +1,6 @@
+export const EVENT_MOUNT = 'mount';
+export const eventNames = [
+  EVENT_MOUNT
+];
+
+export const isEventName = name => eventNames.indexOf(name) !== -1;