+++ src/main/javascript/Core/EventDispatcher.js
@@ -1,0 +1,3 @@
+import { EventEmitter } from 'eventemitter3';
+export const RequestEventDispatcher = new EventEmitter();
+export const ResponseEventDispatcher = new EventEmitter();