+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+* [MAINTENANCE] use eventemitter for async communications
+
 ## v0.2.1 - 2017-05-02
 
 * [FIX] remove hardcoded child window url